Build a Film Club Web App on Google AppEngine/Film Club App/requirements.txt
Flask==1.1.2
google-cloud-datastore==1.12.0
google-auth==1.17.2
requests==2.24.0
